MachineTranslation-Transformer/Tensorflow/source/requirements.gpu-cu80.txt
numba
tensorflow-gpu==1.3.0
sympy
bz2file
